+++ src/monitored_item/delete_monitored_items.rs
@@ -56,11 +56,10 @@ unsafe extern "C" fn callback_execute_response_c(
     _client: *mut UA_Client,
     userdata: *mut c_void,
     _request_id: UA_UInt32,
-    response: *mut c_void,
+    response: *mut UA_DeleteMonitoredItemsResponse,
 ) {
     log::debug!("MonitoredItems_delete_async() completed");
 
-    let response = response.cast::<UA_DeleteMonitoredItemsResponse>();
     // SAFETY: Incoming pointer is valid for access.
     // PANIC: We expect pointer to be valid when good.
     let response = unsafe { response.as_ref() }.expect("response should be set");
@@ -114,11 +113,10 @@ unsafe extern "C" fn callback_log_response_c(
     _client: *mut UA_Client,
     _userdata: *mut c_void,
     _request_id: UA_UInt32,
-    response: *mut c_void,
+    response: *mut UA_DeleteMonitoredItemsResponse,
 ) {
     log::debug!("MonitoredItems_delete_async() completed");
 
-    let response = response.cast::<UA_DeleteMonitoredItemsResponse>();
     // SAFETY: Incoming pointer is valid for access.
     // PANIC: We expect pointer to be valid when good.
     let response = unsafe { response.as_ref() }.expect("response should be set");