+++ CHANGELOG.md
@@ -11,6 +11,14 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add method `ua::VariableAttributes::with_array_dimensions()`.
 - Add method `Server::add_data_type_node()` and type `DataTypeNode`.
+- Add method `ua::CertificateVerification::memory_store()` and `ua::TrustListDataType`.
+
+### Changed
+
+- Breaking: Upgrade to open62541 version
+  [1.5.0](https://github.com/open62541/open62541/releases/tag/v1.5.0).
+- Breaking: Remove method `ua::CertificateVerification::custom()` and trait
+  `CustomCertificateVerification`.
 
 ## [0.10.1] - 2025-10-29